BIP13 implemented #84
Also implemented regular P2PKH addresses.
New CLI commands for generating P2PKH and P2SH addresses as well as generating a new private key, obtaining a public key from a secret key and decoding a script into assembly.
