+++ src/bitcoin-utility/BitcoinUtility.swift
@@ -5,5 +5,5 @@ struct BitcoinUtility: ParsableCommand {
         commandName: "bcutil",
         abstract: "An all-purpose Bitcoin Utility.",
         version: "1.0.0",
-        subcommands: [Seed.self, HDNew.self, HDToPublic.self, HDPrivate.self, HDPublic.self, MnemonicNew.self, MnemonicToSeed.self])
+        subcommands: [Seed.self, ECNew.self, ECToPublic.self, ECToAddress.self, ScriptToAddress.self, ScriptDecode.self, HDNew.self, HDToPublic.self, HDPrivate.self, HDPublic.self, MnemonicNew.self, MnemonicToSeed.self])
 }
