+++ src/bitcoin-utility/SigVersion.swift
@@ -1,0 +1,4 @@
+import ArgumentParser
+import Bitcoin
+
+extension SigVersion: Decodable, ExpressibleByArgument { }
