+++ src/bitcoin-utility/WalletNetwork.swift
@@ -1,0 +1,4 @@
+import ArgumentParser
+import Bitcoin
+
+extension WalletNetwork: Decodable, ExpressibleByArgument { }
