+++ src/bitcoin-utility/commands/ECNew.swift
@@ -1,0 +1,15 @@
+import ArgumentParser
+import Bitcoin
+import Foundation
+
+/// Generates a new secret key suitable for ECDSA signing.
+struct ECNew: ParsableCommand {
+
+    static var configuration = CommandConfiguration(
+        abstract: "Generates a new secret key suitable for ECDSA signing."
+    )
+
+    mutating func run() throws {
+        print(Wallet.createSecretKey())
+    }
+}
