+++ src/bitcoin-utility/commands/HDNew.swift
@@ -2,14 +2,14 @@ import ArgumentParser
 import Bitcoin
 import Foundation
 
-/// New HD wallet master key from seed.
+/// Generates a new HD wallet master key from seed.
 struct HDNew: ParsableCommand {
 
     static var configuration = CommandConfiguration(
-        abstract: "New HD wallet master key from seed."
+        abstract: "Generates a new HD wallet master key from seed."
     )
 
-    @Argument(help: "Entropy in hexadecimal format.")
+    @Argument(help: "The entropy in hexadecimal format.")
     var seed: String
 
     mutating func run() throws {
