+++ src/bitcoin-utility/commands/HDPrivate.swift
