+++ src/bitcoin-utility/commands/HDPublic.swift
