+++ src/bitcoin-utility/commands/MnemonicNew.swift
