+++ src/bitcoin-utility/commands/MnemonicToSeed.swift
