+++ src/bitcoin-utility/commands/Seed.swift
